Add support for loading files directly out of game archives
This adds a dependency on libgamearchive but finally means files can be edited
directly within archive files without manually extracting them first or merging
them later.

This also introduces the EFailure exception which simplifies a lot of the popup
error messages as the popups can now be shown in only a couple of places.

This also consolidates the supplementary file mapping between SuppMap and
SuppData, so that it's now done by a function, reducing duplicated code.