Add mechanism to edit paths in map editor, and provide hints in the status bar
The virtual layer idea has been reworked to include the viewport and paths as
two virtual layers, with the ability to easily add more.

The status bar code has also been updated to provide room for hints about what's
going on as well as which keys can be pressed in the current state.