Rewrite audio processing and music player
The music player code is now removed entirely, and the libgamemusic player is
used instead.  The PortAudio code is massively cleaned up, and it is now
buffered correctly with latency compensation to keep the GUI sync'd to the
audio.  The playback position is now correctly passed across thread boundaries
so the music editor scrolls events during playback properly now.